# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestió de Magatzem",
        "description": "Documentació Swagger del projecte final",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/proveidors": {
            "get": {
                "tags": [
                    "Proveidors"
                ],
                "summary": "Llistar tots els proveidors",
                "operationId": "f262364cf46aee024f7172e9c5ba536f",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Proveidors"
                ],
                "summary": "Crear un proveidor",
                "operationId": "d564bab79c66a2a5de0291b35e50136e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 150
                                    },
                                    "telefon": {
                                        "type": "string",
                                        "maxLength": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveidor creat"
                    },
                    "422": {
                        "description": "Validació fallida"
                    }
                }
            }
        },
        "/api/proveidors/{id}": {
            "get": {
                "tags": [
                    "Proveidors"
                ],
                "summary": "Obtenir un proveidor per ID",
                "operationId": "be996cf1b6c5d4963a29313071f6f128",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveidor trobat"
                    },
                    "404": {
                        "description": "Proveidor no trobat"
                    }
                }
            },
            "put": {
                "tags": [
                    "Proveidors"
                ],
                "summary": "Actualitzar un proveidor",
                "operationId": "13c6a769925dc6535c977d144cd7ef40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 150
                                    },
                                    "telefon": {
                                        "type": "string",
                                        "maxLength": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proveidor actualitzat"
                    },
                    "404": {
                        "description": "Proveidor no trobat"
                    },
                    "422": {
                        "description": "Validació fallida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Proveidors"
                ],
                "summary": "Eliminar un proveidor",
                "operationId": "6b0387d9b1a962360e1117623c007791",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveidor eliminat"
                    },
                    "404": {
                        "description": "Proveidor no trobat"
                    }
                }
            }
        }
    },
...
}